fix: address review comment https://github.com/pcamarajr/rustalk/pull/39#discussion_r2543453065
